refactor(operator_commands): split tests_dispatch.rs (#1266)
tests_dispatch.rs 453 -> 186 LOC; new tests_dispatch_extra.rs.
